+++ packages/react/components/Tabs/index.ts
@@ -1,0 +1,2 @@
+export * from "./Component";
+export * from "./Hook";
